still working to fix the erreur

+++ srcs/requirements/wordpress/conf/auto_config.sh
@@ -1,22 +1,23 @@
 #!/bin/bash
 	mkdir -p /run/php/;
-	touch /run/php/php7.3-fpm.pid;
+	touch /run/php/php7.3-fpm.pid; #Store PID files for PHP processes managed by the PHP-FPM
 	sed -i "s/listen = \/run\/php\/php7.3-fpm.sock/listen = 9000/" "/etc/php/7.3/fpm/pool.d/www.conf"
-	apt-get -y install wget
-	wget https://fr.wordpress.org/wordpress-6.3.1-en_GB.tar.gz -P /var/www
-	cd /var/www 
-	tar -xzf wordpress-6.3.1-en_GB.tar.gz 
-	rm wordpress-6.3.1-en_GB.tar.gz
-	chmod -x /var/www/wordpress
- 	wget https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+ if [ ! -f /var/www/html/wp-config.php ]; then
+	mkdir -p /var/www/html
+	wget https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar; #download wp-cli (WordPress Command Line Interface pour managing your site)
 	chmod +x wp-cli.phar; 
-	mv wp-cli.phar /usr/local/bin/wp; 
-	cd /var/www/wordpress
-	touch /var/www/wordpress/wp-config.php
-	sed -i "s/database_name_here/${SQL_DATABASE}/g" "/var/www/wordpress/wp-config-sample.php"
-	sed -i "s/username_here/${SQL_USER}/g" "/var/www/wordpress/wp-config-sample.php"
-	sed -i "s/password_here/${SQL_PASSWORD}/g" "/var/www/wordpress/wp-config-sample.php"
-	sed -i "s/localhost/${db_host}/g" "/var/www/wordpress/wp-config-sample.php"
-	cp /var/www/wordpress/wp-config-sample.php /var/www/wordpress/wp-config.php
-	echo "success!"
+	mv wp-cli.phar /usr/local/bin/wp; # you might use the wp plugin install command to install a plugin, or the wp theme list command to list the installed themes.
+	cd /var/www/html;
+	wp core download --allow-root; #download the latest version of WordPress  and The --allow-root option allows the command to be executed by the root user
+	touch /var/www/html/wp-config.php;
+	sed -i "s/database_name_here/${SQL_DATABASE}/g" "/var/www/html/wp-config-sample.php"
+	sed -i "s/username_here/${SQL_USER}/g" "/var/www/html/wp-config-sample.php"
+	sed -i "s/password_here/${SQL_PASSWORD}/g" "/var/www/html/wp-config-sample.php"
+	sed -i "s/localhost/${db_host}/g" "/var/www/html/wp-config-sample.php"
+	cp /var/www/html/wp-config-sample.php /var/www/html/wp-config.php;
+	# install WordPress using the specified options
+	wp core install --allow-root --url=$URL --title=$TITLE --admin_user=$ADMIN_USER --admin_password=$ADMIN_PSSWRD --admin_email=$ADMIN_EMAIL
+	wp user create --allow-root ${SQL_USER} ${USER_EMAIL} --user_pass=${SQL_PASSWORD}; # create a new user for your WordPress site
+	echo "successful !"
+ fi
 exec "$@"
